Relocate Template Menu to common components
