--- components/common/BoardEditor/focalboard/src/components/viewHeader/TemplatesMenu/index.ts
@@ -1,1 +1,0 @@
-export { TemplatesMenu } from './TemplatesMenu'
