+++ components/common/TemplatesMenu/index.ts
@@ -1,0 +1,1 @@
+export { TemplatesMenu } from './TemplatesMenu';
